HDDS-2898. build failure due to hadoop-hdds-client test
